Week 2/DAY 5/pgsqlCopyEditbook_review_api/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.17
pydantic==2.0.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
psycopg2-binary==2.9.6